requirements.txt
mss>=9.0.0
pyautogui>=0.9.54
pygetwindow>=0.0.9
gspread>=6.0.0
google-auth>=2.0.0
google-genai>=1.0.0
Pillow>=10.0.0
streamlit>=1.30.0
pandas>=2.0.0
